ml-service/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
